refactor: createOmegaQuote
