+++ src/context/paging/omegaquotesPaging.tsx
@@ -2,7 +2,7 @@
 import generatePagingProvider, { generatePagingContext } from '@/context/paging/PagingGenerator'
 import { readQuotesPage } from '@/actions/omegaquotes/read'
 import type { ReadPageInput } from '@/actions/Types'
-import type { OmegaquoteFiltered } from '@/actions/omegaquotes/Types'
+import type { OmegaquoteFiltered } from '@/server/omegaquotes/Types'
 
 export type PageSizeOmegaquote = 20;
 const fetcher = async (x: ReadPageInput<PageSizeOmegaquote>) => await readQuotesPage(x)
