+++ src/server/omegaquotes/Types.ts
